# openapi.yaml
openapi: 3.0.0
paths:
...
  /admin/profile:
    post:
      operationId: AdminProfileController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminResponse"
      summary: Create an admin (returns existing if email exists)
      tags: &a3
        - Admin
  /admin/profile/{id}:
    get:
      operationId: AdminProfileController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminResponse"
      summary: Get an admin profile by id
      tags: *a3
    patch:
      operationId: AdminProfileController_updateProfile
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminProfileDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminResponse"
      summary: Update an admin profile
      tags: *a3
...
  /customers:
    post:
      operationId: CustomerController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
      security: &a8
        - bearer: []
      summary: Create a customer
      tags: &a9
        - Customers
    get:
      operationId: CustomerController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: q
          required: false
          in: query
          description: Free-text search query.
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCustomerResponse"
      security: *a8
      summary: List customers (paginated, searchable)
      tags: *a9
  /customers/import:
    post:
      operationId: CustomerController_importCustomers
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportCustomersDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResultResponse"
      security: *a8
      summary: Bulk import customers
      tags: *a9
  /customers/{id}:
    get:
      operationId: CustomerController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
      security: *a8
      summary: Get a customer by id
      tags: *a9
    patch:
      operationId: CustomerController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
      security: *a8
      summary: Update a customer by id
      tags: *a9
    delete:
      operationId: CustomerController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
      security: *a8
      summary: Delete a customer by id
      tags: *a9
  /appointments:
    post:
      operationId: AppointmentController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
      security: &a10
        - bearer: []
      summary: Create an appointment
      tags: &a11
        - Appointments
    get:
      operationId: AppointmentController_list
      parameters:
        - name: from
          required: true
          in: query
          schema:
            type: string
        - name: to
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentResponse"
      security: *a10
      summary: List appointments
      tags: *a11
  /appointments/{id}:
    get:
      operationId: AppointmentController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
      security: *a10
      summary: Get an appointment by id
      tags: *a11
    patch:
      operationId: AppointmentController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
      security: *a10
      summary: Update an appointment
      tags: *a11
    delete:
      operationId: AppointmentController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
      security: *a10
      summary: Delete an appointment
      tags: *a11
  /appointments/{id}/status:
    patch:
      operationId: AppointmentController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentStatusDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
      security: *a10
      summary: Update an appointment's status
      tags: *a11
...
info:
  title: Ventura API
  description: Ventura backend API.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: JWT access token from /auth sign-in.
  schemas:
...
    CreateAdminDto:
      type: object
      properties:
        name:
          type: string
          example: Ada Lovelace
        email:
          type: string
          example: ada@example.com
          format: email
      required:
        - name
        - email
    AdminResponse:
      type: object
      properties:
        id:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        shortId:
          type: string
          example: aB3xY9kP
        name:
          type: string
          example: Ada Lovelace
        email:
          type: string
          example: ada@example.com
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - shortId
        - name
        - email
        - createdAt
        - updatedAt
    UpdateAdminProfileDto:
      type: object
      properties:
        name:
          type: string
          example: Ada Lovelace
...
    CreateCustomerDto:
      type: object
      properties:
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com
        phone:
          type: string
          example: +1-555-987-6543
        notes:
          type: string
          example: Prefers email contact.
      required:
        - name
    CustomerResponse:
      type: object
      properties:
        id:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        shortId:
          type: string
          example: aB3xY9kP
        businessId:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        name:
          type: string
          example: Jane Doe
        email:
          type: object
          nullable: true
          format: email
          example: jane@example.com
        phone:
          type: object
          nullable: true
          example: +1-555-987-6543
        notes:
          type: object
          nullable: true
          example: Prefers email contact.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - shortId
        - businessId
        - name
        - createdAt
        - updatedAt
    ImportCustomersDto:
      type: object
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/CreateCustomerDto"
      required:
        - customers
    BulkImportResultResponse:
      type: object
      properties:
        created:
          type: array
          items:
            $ref: "#/components/schemas/CustomerResponse"
        skipped:
          example:
            - index: 2
              reason: Duplicate email.
          type: array
          items:
            type: object
        failed:
          example:
            - index: 5
              reason: Validation failed.
          type: array
          items:
            type: object
      required:
        - created
        - skipped
        - failed
    PaginationMeta:
      type: object
      properties:
        total:
          type: number
          example: 137
        page:
          type: number
          example: 1
        limit:
          type: number
          example: 20
        totalPages:
          type: number
          example: 7
      required:
        - total
        - page
        - limit
        - totalPages
    PaginatedCustomerResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CustomerResponse"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    UpdateCustomerDto:
      type: object
      properties:
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com
        phone:
          type: string
          example: +1-555-987-6543
        notes:
          type: string
          example: Prefers email contact.
    InviteeDto:
      type: object
      properties:
        name:
          type: string
          example: Ada Lovelace
        email:
          type: string
          example: ada@example.com
          format: email
        customerId:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
      required:
        - name
    RecurrenceDto:
      type: object
      properties:
        frequency:
          type: string
          enum:
            - daily
            - weekly
            - monthly
        interval:
          type: number
          example: 1
        until:
          type: string
          example: 2026-12-31T00:00:00.000Z
      required:
        - frequency
    CreateAppointmentDto:
      type: object
      properties:
        title:
          type: string
          example: Client consultation
        start:
          type: string
          example: 2026-06-17T09:00:00.000Z
        end:
          type: string
          example: 2026-06-17T10:00:00.000Z
        notes:
          type: string
          example: Bring the signed contract.
        location:
          type: string
          example: Office, Accra
        invitees:
          type: array
          items:
            $ref: "#/components/schemas/InviteeDto"
        recurrence:
          $ref: "#/components/schemas/RecurrenceDto"
      required:
        - title
        - start
        - end
    InviteeResponse:
      type: object
      properties:
        name:
          type: string
          example: Ada Lovelace
        email:
          type: object
          nullable: true
          example: ada@example.com
        customerId:
          type: object
          nullable: true
          example: 665f1b2c3d4e5f6a7b8c9d0e
      required:
        - name
    RecurrenceResponse:
      type: object
      properties:
        frequency:
          type: string
          enum:
            - daily
            - weekly
            - monthly
        interval:
          type: number
          example: 1
        until:
          type: string
          nullable: true
          format: date-time
      required:
        - frequency
        - interval
    AppointmentResponse:
      type: object
      properties:
        id:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        shortId:
          type: string
          example: aB3xY9kP
        businessId:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        createdBy:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        title:
          type: string
          example: Client consultation
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        notes:
          type: object
          nullable: true
        location:
          type: object
          nullable: true
          example: Office, Accra
        invitees:
          type: array
          items:
            $ref: "#/components/schemas/InviteeResponse"
        recurrence:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/RecurrenceResponse"
        status:
          type: string
          enum:
            - scheduled
            - completed
            - attended
            - cancelled
          example: scheduled
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - shortId
        - businessId
        - createdBy
        - title
        - start
        - end
        - invitees
        - status
        - createdAt
        - updatedAt
    UpdateAppointmentDto:
      type: object
      properties:
        title:
          type: string
          example: Client consultation
        start:
          type: string
          example: 2026-06-17T09:00:00.000Z
        end:
          type: string
          example: 2026-06-17T10:00:00.000Z
        notes:
          type: string
          example: Bring the signed contract.
        location:
          type: string
          example: Office, Accra
        invitees:
          type: array
          items:
            $ref: "#/components/schemas/InviteeDto"
        recurrence:
          $ref: "#/components/schemas/RecurrenceDto"
        clearRecurrence:
          type: boolean
          example: true
    UpdateAppointmentStatusDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - scheduled
            - completed
            - attended
            - cancelled
          example: completed
      required:
        - status
...