# src/admin/controllers/admin-profile.controller.ts
import {
  Body,
  Controller,
  Get,
  HttpCode,
  HttpStatus,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
import { AdminProfileService } from '../application/admin-profile.service';
import { toAdminResponse } from '../application/admin.mapper';
import {
  CreateAdminDto,
  UpdateAdminProfileDto,
} from '../dto/admin-profile.dto';
import { AdminResponse } from '../responses/admin.response';

@ApiTags('Admin')
@Controller('admin/profile')
export class AdminProfileController {
  constructor(private readonly adminProfileService: AdminProfileService) {}

  @ApiOperation({
    summary: 'Create an admin (returns existing if email exists)',
  })
  @ApiResponse({ status: 200, type: AdminResponse })
  @HttpCode(HttpStatus.OK)
  @Post()
  async create(@Body() dto: CreateAdminDto) {
    return toAdminResponse(await this.adminProfileService.create(dto));
  }

  @ApiOperation({ summary: 'Get an admin profile by id' })
  @ApiResponse({ status: 200, type: AdminResponse })
  @Get('/:id')
  async getById(@Param('id') id: string) {
    return toAdminResponse(await this.adminProfileService.getById(id));
  }

  @ApiOperation({ summary: 'Update an admin profile' })
  @ApiResponse({ status: 200, type: AdminResponse })
  @HttpCode(HttpStatus.OK)
  @Patch('/:id')
  async updateProfile(
    @Param('id') id: string,
    @Body() dto: UpdateAdminProfileDto,
  ) {
    return toAdminResponse(
      await this.adminProfileService.updateProfile(id, dto),
    );
  }
}