# src/appointment/dto/update-appointment-status.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEnum } from 'class-validator';
import { AppointmentStatus } from '../domain/appointment.entity';

/** Move an appointment to a new lifecycle status. */
export class UpdateAppointmentStatusDto {
  @ApiProperty({
    enum: AppointmentStatus,
    example: AppointmentStatus.COMPLETED,
  })
  @IsEnum(AppointmentStatus)
  status!: AppointmentStatus;
}